Add screen home and fix cart payment

+++ marketplace/distance.py
@@ -20,14 +20,14 @@ def cost_distance_km(distance_km):
     if distance_km is None:
         return 15000
     if isinstance(distance_km, str):
-        distance_km = float(distance_km.split()[0])
+        distance_km = int(distance_km.split()[0])
 
     if distance_km <= 3.0:
-        return FIRST_KM_COST
+        return int(FIRST_KM_COST)
     else:
         # Tính phí cho 3 km đầu tiên
         total_cost = FIRST_KM_COST
         # Tính phí cho các km vượt quá
         excess_distance = distance_km - 3.0
         total_cost += excess_distance * LAST_KM_COST
-        return total_cost
+        return int(total_cost)
